Add cost to techs in code, change fitness to float
also add fitnessGain bonus to hunt and collect actions

+++ src/caveman.cpp
@@ -50,10 +50,10 @@ void Caveman::aging() {
     age++;
 }
 
-short Caveman::getFitness(){
+float Caveman::getFitness(){
     return fitness;
 }
-void Caveman::setFitness(short newFit) {
+void Caveman::setFitness(float newFit) {
     fitness = newFit;
 }
 short Caveman::getIntelligence(){
